+++ notify.c
@@ -233,10 +233,31 @@ fdm_notify_stdout(struct fdm *fdm, int fd, int events, void *data)
     return true;
 }
 
+struct done_context {
+    struct wayland *wayl;
+    struct terminal *term;
+};
+
 static void
 notif_done(struct reaper *reaper, pid_t pid, int status, void *data)
 {
-    struct terminal *term = data;
+    struct done_context *ctx = data;
+    const struct wayland *wayl = ctx->wayl;
+    const struct terminal *wanted_term = ctx->term;
+    free(ctx);
+
+    struct terminal *term = NULL;
+    tll_foreach(wayl->terms, it) {
+        if (it->item == wanted_term) {
+            term = it->item;
+            break;
+        }
+    }
+
+    if (term == NULL) {
+        LOG_WARN("notification closed, but the associated terminal instance is already gone");
+        return;
+    }
 
     tll_foreach(term->active_notifications, it) {
         struct notification *notif = &it->item;
@@ -567,11 +588,18 @@ notify_notify(struct terminal *term, struct notification *notif)
                 &fdm_notify_stdout, (void *)term);
     }
 
+    struct done_context *ctx = NULL;
+    if (track_notification) {
+        ctx = xmalloc(sizeof(*ctx));
+        ctx->wayl = term->wl;
+        ctx->term = term;
+    }
+
     /* Redirect stdin to /dev/null, but ignore failure to open */
     int devnull = open("/dev/null", O_RDONLY);
     pid_t pid = spawn(
         term->reaper, NULL, argv, devnull, stdout_fds[1], -1,
-        track_notification ? &notif_done : NULL, (void *)term, NULL);
+        track_notification ? &notif_done : NULL, ctx, NULL);
 
     if (stdout_fds[1] >= 0) {
         /* Close write-end of stdout pipe */
@@ -649,9 +677,7 @@ notify_close(struct terminal *term, const char *id)
             }
 
             int devnull = open("/dev/null", O_RDONLY);
-            spawn(
-                term->reaper, NULL, argv, devnull, -1, -1,
-                NULL, (void *)term, NULL);
+            spawn(term->reaper, NULL, argv, devnull, -1, -1, NULL, NULL, NULL);
 
             if (devnull >= 0)
                 close(devnull);