+++ CHANGELOG.md
@@ -114,6 +114,9 @@
   ([#2364][2364]).
 * Escape quotes in file names being DnD:ed on the command line
   ([#2363][2363]).
+* Crash in `--server` mode, when a tracked notification is closed
+  after the associated terminal instance has been closed
+  ([#2397][2397]).
 
 [2353]: https://codeberg.org/dnkl/foot/issues/2353
 [2352]: https://codeberg.org/dnkl/foot/issues/2352
@@ -122,6 +125,7 @@
 [2360]: https://codeberg.org/dnkl/foot/issues/2360
 [2370]: https://codeberg.org/dnkl/foot/issues/2370
 [2364]: https://codeberg.org/dnkl/foot/issues/2364
+[2397]: https://codeberg.org/dnkl/foot/issues/2397
 
 
 ### Security